src/requirements.txt
Flask==1.1.1
pymongo==3.9.0
flasgger==0.9.2
python-dateutil==2.8.1
flask-cors==3.0.8
pyjwt